requirements.txt
langchain==0.3.12
langchain_community==0.3.12
langchain_core==0.3.25
langchain_openai==0.2.12
matplotlib==2.1.0
numpy==1.13.3
PyYAML==6.0.2
transformers==4.47.0
rouge_score==0.1.2
autoawq==0.2.7.post3
jinja2==3.1.0